build: Add spdlog build dependency
spdlog is included and built from source, using the third_party/spdlog
submodule.

When Gloo is built as a static library, spdlog is used in its
header-only library form. When Gloo is built as a shared library, spdlog
is built statically and included.

A key goal is to make spdlog a dependency internal to Gloo only. Gloo
users should not have to install spdlog themselves.

This commit does not add any usages of spdlog in Gloo, just the build
infrastructure to use it.